refactor: 💡 updates based on PR comments
