+++ src/components/DataRoom/TransactionsOnChain/utils.ts
@@ -1,37 +1,16 @@
-import type { MotionValue } from 'framer-motion'
-import { useEffect, useState } from 'react'
+// This function calculates the Y position for individual digits in the counter animation.
+// It creates a smooth rolling effect from 0 to the targetNumber.
 
-export function calculateYPosition(latest: number, number: number, height: number): number {
-  let placeValue = latest % 10
-  let offset = (10 + number - placeValue) % 10
-  let memo = offset * height
-  if (offset > 5) {
-    memo -= 10 * height
+export function calculateYPosition(latest: number, targetNumber: number, digitHeight: number): number {
+  const currentDigit = latest % 10
+  const digitDifference = (10 + targetNumber - currentDigit) % 10
+
+  let yOffset = digitDifference * digitHeight
+
+  // Adjust for shortest path (up or down)
+  if (digitDifference > 5) {
+    yOffset -= 10 * digitHeight
   }
-  return memo
-}
 
-// Custom hook for managing counter based on scroll
-export function useCounterScroll(scrollYProgress: MotionValue<number>, percentage: number): number {
-  const [value, setValue] = useState(0)
-
-  useEffect(() => {
-    const checkScrollProgress = () => {
-      const latest = scrollYProgress.get()
-      if (latest >= 0.11) {
-        setValue(percentage)
-      } else if (latest <= 0.1) {
-        setValue(0)
-      }
-    }
-
-    const unsubscribe = scrollYProgress.on('change', checkScrollProgress)
-
-    // Initial check
-    checkScrollProgress()
-
-    return () => unsubscribe()
-  }, [scrollYProgress, percentage])
-
-  return value
+  return yOffset
 }
